+++ TestBench/TestBench.cpp
@@ -1,10 +1,14 @@
 
 #include <stdafx.h>
-// From https://github.com/pytorch/examples/blob/master/cpp/transfer-learning/main.cpp
+
+// reference example code:  https://github.com/pytorch/examples/blob/master/cpp/transfer-learning/main.cpp
+
 #include <memory>
 #include <algorithm>
 #include <vector>
 #include <fstream>
+#include <iomanip>
+#include <thread>
 
 #include "CLI/CLI.hpp"
 #include <torch/torch.h>
@@ -12,6 +16,7 @@
 #include "torchvision/models/resnet.h"
 #include "opencv2/opencv.hpp"
 #include "opencv2/imgcodecs.hpp"
+#include <opencv2/core/utils/filesystem.hpp>
 
 #include "data/DataAugmentation.h"
 #include "nn/models/EfficientNet.h"
@@ -107,50 +112,107 @@ int main(int argc, const char* argv[]) {
         app.allow_windows_style_options();
         #endif
 
-        ArchType archtype { ArchType::ResNet34 };
-        app.add_option("-a,--arch", archtype, "Architecture")
-            ->default_val(ArchType::ResNet34)
-            ->transform(CLI::CheckedTransformer(ArchMap, CLI::ignore_case));
-        OptimType optimtype { OptimType::SGD };
-        app.add_option("-o,--optim", optimtype, "Optimizer")
-            ->default_val(OptimType::SGD)
-            ->transform(CLI::CheckedTransformer(OptimMap, CLI::ignore_case));
-
         std::string dataset_folder_train, dataset_folder_valid;
         std::string data_file_extension = "jpeg";
         std::string log_file_path;
-        double lr{ 0.001 };
-        double clipping{ 0.01 };
+        double lr{ 1e-3 };   // only param that is mandatory to initialize options
+        double clipping{ -1.0 }, dampening{ -1.0 }, momentum{ -1.0 }, epsilon{ -1.0 }, weight_decay{ -1.0 };
+        std::tuple<double, double> betas{ -1, -1 };
+        bool nesterov, amsgrad;
         size_t batch_size = 16;
         size_t max_epochs = 30;
+        size_t max_train_samples = 0, max_valid_samples = 0;
+        unsigned int seed = unsigned(std::time(nullptr));  // always random if not overridden
         int workers = -1;  // allow undefined
         bool debug{ false };
         bool verbose{ false };
         bool version{ false };
-        bool no_color{ false };
+        bool log_color{ true };
 
-        app.add_option("--train", dataset_folder_train,
+        auto train_opts = app.add_option_group("Training Parameters",
+            "Parameters that define behavior of training iterations and data loading.");
+        train_opts->add_option("--train", dataset_folder_train,
             "Directory where training images categorized by class sub-folders can be loaded from."
-        );
-        app.add_option("--valid", dataset_folder_valid,
+        )->check(CLI::ExistingDirectory)->required();
+        train_opts->add_option("--valid", dataset_folder_valid,
             "Directory where validation images categorized by class sub-folders can be loaded from."
-        );
-        app.add_option("-e,--extension", data_file_extension,
+        )->check(CLI::ExistingDirectory)->required();
+        train_opts->add_option("-e,--extension", data_file_extension,
             "Extension of image files to be considered for loading data."
         )->default_val(data_file_extension);
+        train_opts->add_option("--max-train-samples", max_train_samples,
+            "Maximum amount of samples to preserve from available training dataset directory. "
+            "Consume all available samples if not specified, 0 or greater than available amount."
+        )->check(CLI::PositiveNumber);
+        train_opts->add_option("--max-valid-samples", max_valid_samples,
+            "Maximum amount of samples to preserve from available validation dataset directory. "
+            "Consume all available samples if not specified, 0 or greater than available amount."
+        )->check(CLI::PositiveNumber);
+        train_opts->add_option("-E,--max-epochs", max_epochs,
+            "Maximum number of training epochs."
+        )->check(CLI::PositiveNumber)->default_val(max_epochs);
+        train_opts->add_option("-B,--batch-size", batch_size,
+            "Batch size of each iteration."
+        )->check(CLI::PositiveNumber)->default_val(batch_size);
+        train_opts->add_option("-W,--workers", workers,
+            "Number of data loader workers to employ for data loading. "
+            "If not specified or negative (default), employ whichever amount available based on CPU count. "
+            "If equal to zero, run in single-thread mode (worker will be on the same thread as main process). "
+            "Otherwise, a positive integer indicates explicitly how many worker threads to employ."
+        )->default_val(workers);
+        train_opts->add_option("-S,--seed", seed,
+            "Seed for initialization of random number generator applied wherever needed."
+        )->check(CLI::PositiveNumber);
 
-        app.add_option("-E,--max-epochs", max_epochs, "Maximum number of training epochs.")->default_val(max_epochs);
-        app.add_option("-B,--batch-size", batch_size, "Batch size of each iteration.")->default_val(batch_size);
-        app.add_option("-W,--workers", workers, "Number of data loader workers to employ.")->default_val(workers);
-        CLI::Option* lr_opt = app.add_option("--lr", lr, "Learning rate")->default_val(lr);
-        CLI::Option* clipping_opt = app.add_option("--clipping", clipping, "Clipping threshold")->default_val(clipping);
+        auto model_opts = app.add_option_group("Model Parameters",
+            "Parameters relevant for model selection and configuration.");
+        ArchType arch_type { ArchType::ResNet34 };
+        model_opts->add_option("-a,--arch", arch_type, "Architecture")
+            ->default_val(arch_type)
+            ->transform(CLI::CheckedTransformer(ArchMap, CLI::ignore_case));
+        std::string ckpt_load_path, ckpt_save_path;
+        model_opts->add_option("-c,--checkpoint", ckpt_load_path,
+            "Model checkpoint to load (must match architecture)."
+        )->check(CLI::ExistingFile);
+        model_opts->add_option("-s,--save-dir", ckpt_save_path,
+            "Save location of intermediate epoch model checkpoints."
+        )->default_val("./checkpoints");  // doesn't need to exist, will create if missing
 
-        app.add_option("-l,--logfile", log_file_path, "Output log path.");
-        app.add_flag("-d,--debug", debug, "Log additional debug messages.")->default_val(debug);
-        app.add_flag("-v,--verbose", verbose,
+        // because many different optimizers use the same hyperparameter names for different purposes / defaults,
+        // don't define any default here and use CLI option to detect if specific OptimizerOption default must be used
+        auto optim_opts = app.add_option_group("Optimizer Hyperparameters",
+            "Hyperparameters values employed by the selected optimizer. "
+            "Omitted options use PyTorch defaults. Not applicable options for selected optimizer are ignored.");
+        OptimType optim_type { OptimType::SGD };
+        optim_opts->add_option("-o,--optim", optim_type, "Optimizer")
+            ->default_val(optim_type)
+            ->transform(CLI::CheckedTransformer(OptimMap, CLI::ignore_case));
+        auto lr_opt = optim_opts->add_option("--lr", lr, "Learning rate")->check(CLI::PositiveNumber);
+        auto epsilon_opt = optim_opts->add_option("--epsilon", epsilon, "Epsilon")->check(CLI::NonNegativeNumber);
+        auto amsgrad_opt = optim_opts->add_option("--amsgrad", amsgrad, "AMSGrad");
+        auto weight_decay_opt = optim_opts->add_option("--weight-decay", amsgrad, "Weight Decay")->check(CLI::NonNegativeNumber);
+        auto clipping_opt = optim_opts->add_option("--clipping", clipping, "Clipping lambda threshold")->check(CLI::NonNegativeNumber);
+        auto momentum_opt = optim_opts->add_option("--momentum", clipping, "Momentum")->check(CLI::NonNegativeNumber);
+        auto dampening_opt = optim_opts->add_option("--dampening", clipping, "Dampening")->check(CLI::NonNegativeNumber);
+        auto nesterov_opt = optim_opts->add_option("--nesterov", nesterov, "Nesterov");
+        auto betas_opt = optim_opts->add_option("--betas", betas,
+            "Beta1 and Beta2 (both requried if option is specified, separated by space)."
+        )->check(CLI::PositiveNumber);/*->expected(2) */  // expected size automatic with tuple
+
+        #ifndef USE_LOG_COUT
+
+        auto log_opts = app.add_option_group("Logging Parameters", "Control logging options.");
+        log_opts->add_flag("-v,--verbose", verbose,
             "Log more verbose messages including function and lines numbers."
         )->default_val(verbose);
-        app.add_flag("--no-color", no_color, "Disable color formatting of log entries.")->default_val(no_color);
+        log_opts->add_flag("-d,--debug", debug, "Log additional debug messages.")->default_val(debug);
+        log_opts->add_option("-l,--logfile", log_file_path, "Output log path.");
+        log_opts->add_flag("--color,!--no-color", log_color,
+            "Enable/disable color formatting of log entries."
+        )->default_val(log_color);
+
+        #endif // USE_LOG_COUT
+
         app.add_flag("-V,--version", version, "Print the version number.");
 
         CLI11_PARSE(app, argc, argv);
@@ -159,6 +221,7 @@ int main(int argc, const char* argv[]) {
         auto start_time = std::chrono::steady_clock::now();
 
         if (version) {
+            // cout to force output regardless of log level / log utility
             std::cout << std::string(CRIM_TORCH_EXTENSIONS_VERSION) << std::endl;
             return EXIT_SUCCESS;
         }
@@ -172,57 +235,94 @@ int main(int argc, const char* argv[]) {
                 static plog::ColorConsoleAppender<plog::VerboseFormatter> colorConsoleAppender;
                 static plog::RollingFileAppender<plog::VerboseFormatter> fileAppender(log_file_path.c_str());
                 if (withFile)
-                    plog::init(level, no_color ? &consoleAppender : &colorConsoleAppender).addAppender(&fileAppender);
+                    plog::init(level, log_color ? &colorConsoleAppender : &consoleAppender).addAppender(&fileAppender);
                 else
-                    plog::init(level, no_color ? &consoleAppender : &colorConsoleAppender);
+                    plog::init(level, log_color ? &colorConsoleAppender : &consoleAppender);
             }
             else {
                 static plog::ConsoleAppender<plog::MinimalFormatter> consoleAppender;
                 static plog::ColorConsoleAppender<plog::MinimalFormatter> colorConsoleAppender;
                 static plog::RollingFileAppender<plog::MinimalFormatter> fileAppender(log_file_path.c_str());
                 if (withFile)
-                    plog::init(level, no_color ? &consoleAppender : &colorConsoleAppender).addAppender(&fileAppender);
+                    plog::init(level, log_color ? &colorConsoleAppender : &consoleAppender).addAppender(&fileAppender);
                 else
-                    plog::init(level, no_color ? &consoleAppender : &colorConsoleAppender);
+                    plog::init(level, log_color ? &colorConsoleAppender : &consoleAppender);
             }
 
         #endif // USE_LOG_COUT
 
+        uint w = 20;
+        std::string sep = ": ", tab = "  ";
+        auto worker_cpu = std::thread::hardware_concurrency();
+        size_t worker_count = (workers < 0 ? worker_cpu : workers);
+        LOGGER(INFO) << "Options" << std::endl
+            << tab << std::setw(w) << std::left << "Workers" << sep << worker_count
+                << (workers != worker_cpu ? " (auto)" : "") << std::endl
+            << tab << std::setw(w) << std::left << "Batch Size" << sep << batch_size << std::endl
+            << tab << std::setw(w) << std::left << "Max Epochs" << sep << max_epochs << std::endl;
+
         bool has_cuda = torch::cuda::is_available();
-        if (verbose) {
-            if(lr_opt->count() > 0)
-                LOGGER(DEBUG) << "Learning rate = " << lr << std::endl;
-            if(clipping_opt->count()>0)
-                LOGGER(DEBUG) << "Lambda = " << clipping << std::endl;
-            LOGGER(INFO) << (has_cuda ? "CUDA detected!" : "CUDA missing! Will use CPU.") << std::endl;
-            if (has_cuda)
-                show_gpu_properties();
-        }
+        LOGGER(INFO) << (has_cuda ? "CUDA detected!" : "CUDA missing! Will use CPU.") << std::endl;
+        if (has_cuda)
+            show_gpu_properties();
 
         if (dataset_folder_train.empty() || dataset_folder_valid.empty()) {
             LOGGER(ERROR) << "Invalid directories for train/valid datasets provided no data!" << std::endl;
             return EXIT_FAILURE;
         }
 
-        LOGGER(INFO) << "Loading samples..." << std::endl;
+        if (!ckpt_load_path.empty()) {
+            ckpt_load_path = cv::utils::fs::canonical(ckpt_load_path);
+            LOGGER(INFO) << "Will attempt loading model checkpoint from file: [" << ckpt_load_path << "]" << std::endl;
+        }
+        ckpt_save_path = cv::utils::fs::canonical(ckpt_save_path);
+        if (!cv::utils::fs::isDirectory(ckpt_save_path)) {
+            cv::utils::fs::createDirectory(ckpt_save_path);
+        }
+        LOGGER(DEBUG) << "Will save epoch checkpoints in [" << ckpt_save_path << "]" << std::endl;
+
+        LOGGER(INFO) << "Searching directories for samples..." << std::endl;
 
         // Get paths of images and labels from the folder paths
-        std::pair<std::vector<std::string>, std::vector<Label>> samples_train = load_data_from_folder(
-            dataset_folder_train, data_file_extension
-        );
-        std::pair<std::vector<std::string>, std::vector<Label>> samples_valid = load_data_from_folder(
-            dataset_folder_valid, data_file_extension
-        );
+        DataSamples samples_train = load_data_from_class_folder_tree(dataset_folder_train, data_file_extension, workers);
+        DataSamples samples_valid = load_data_from_class_folder_tree(dataset_folder_valid, data_file_extension, workers);
 
         size_t nb_class_train = count_classes(samples_train.second);
         size_t nb_class_valid = count_classes(samples_valid.second);
         size_t nb_class = std::max(nb_class_train, nb_class_valid);
+        size_t nb_total_train = samples_train.first.size();
+        size_t nb_total_valid = samples_valid.first.size();
+        size_t nb_total = nb_total_train + nb_total_valid;
 
-        LOGGER(INFO) << "Number of found classes: " << nb_class << std::endl;
-        LOGGER(INFO) << "Number of train classes: " << nb_class_train << std::endl;
-        LOGGER(INFO) << "Number of valid classes: " << nb_class_valid << std::endl;
-        LOGGER(INFO) << "Number of train samples: " << samples_train.first.size() << std::endl;
-        LOGGER(INFO) << "Number of valid samples: " << samples_valid.first.size() << std::endl;
+        LOGGER(INFO) << "Number of available total classes: " << nb_class << std::endl;
+        LOGGER(INFO) << "Number of available train classes: " << nb_class_train << std::endl;
+        LOGGER(INFO) << "Number of available valid classes: " << nb_class_valid << std::endl;
+        LOGGER(INFO) << "Number of available total samples: " << nb_total << std::endl;
+        LOGGER(INFO) << "Number of available train samples: " << nb_total_train << std::endl;
+        LOGGER(INFO) << "Number of available valid samples: " << nb_total_valid << std::endl;
+
+        if (max_train_samples != 0 && max_train_samples < nb_total_train) {
+            samples_train = random_pick(samples_train, max_train_samples, seed);
+            nb_class_train = count_classes(samples_train.second);
+            nb_total_train = samples_train.first.size();
+        }
+        if (max_valid_samples != 0 && max_valid_samples < nb_total_train) {
+            samples_valid = random_pick(samples_valid, max_valid_samples, seed);
+            nb_class_valid = count_classes(samples_valid.second);
+            nb_total_valid = samples_valid.first.size();
+        }
+        if (nb_total != nb_total_train + nb_total_valid) {
+            nb_total = nb_total_train + nb_total_valid;
+            nb_class = std::max(nb_class_train, nb_class_valid);
+            LOGGER(WARN) << "Number of samples was modified according to options!" << std::endl;
+            LOGGER(INFO) << "Number of selected total classes: " << nb_class << std::endl;
+            LOGGER(INFO) << "Number of selected train classes: " << nb_class_train << std::endl;
+            LOGGER(INFO) << "Number of selected valid classes: " << nb_class_valid << std::endl;
+            LOGGER(INFO) << "Number of selected total samples: " << nb_total << std::endl;
+            LOGGER(INFO) << "Number of selected train samples: " << nb_total_train << std::endl;
+            LOGGER(INFO) << "Number of selected valid samples: " << nb_total_valid << std::endl;
+        }
+
         show_machine_memory();
         show_gpu_memory();
 
@@ -236,7 +336,7 @@ int main(int argc, const char* argv[]) {
         }
 
         uint64_t image_size = 224;
-        LOGGER(INFO) << "Will resize sample images to:"
+        LOGGER(INFO) << "Will resize sample images to: "
                      << "(" << image_size << ", " << image_size << ") [enforced by model input layer]" << std::endl;
 
         #ifdef USE_BASE_MODEL
@@ -254,24 +354,32 @@ int main(int argc, const char* argv[]) {
 
         //torch::nn::AnyModule pNet;
         std::vector<torch::Tensor> params;
-        switch (archtype) {
+        auto model_loaded = false;
+        std::string load_str = " (loaded: [" + ckpt_load_path + "])";
+        switch (arch_type) {
             case ArchType::ResNet34:
                 {
                     //pNet = vision::models::ResNet34(nb_class);
                     auto p = std::make_shared<ResNet34CLI>(nb_class);
                     #ifdef USE_BASE_MODEL
-                    auto net = p.get();
+                        auto net = p.get();
                     #else
-                    auto net = p;
+                        auto net = p;
+                        if (!ckpt_load_path.empty()) {
+                            torch::load(p, ckpt_load_path);
+                            model_loaded = true;
+                        }
                     #endif
+
                     params = net->parameters();
                     if (has_cuda) net->to(torch::kCUDA);
-                    LOGGER(INFO) << "Using ResNet34" << std::endl;
+                    LOGGER(INFO) << "Using model: ResNet34" << (model_loaded ? load_str : "") << std::endl;
                     LOGGER(DEBUG) << std::endl << *net << std::endl;
+
                     #ifdef USE_BASE_MODEL
-                    pNet = std::dynamic_pointer_cast<IModel>(p);
+                        pNet = std::dynamic_pointer_cast<IModel>(p);
                     #else
-                    pNet = p;
+                        pNet = p;
                     #endif
                     //pNet = torch::nn::AnyModule(ResNet34(nb_class));
                 }
@@ -281,18 +389,24 @@ int main(int argc, const char* argv[]) {
                     //pNet = std::make_shared<EfficientNetV1>(EfficientNetOptions{ 1.0, 1.0, 224, 0.2 }, nb_class);
                     auto p = std::make_shared<EfficientNetB0CLI>(nb_class);
                     #ifdef USE_BASE_MODEL
-                    auto net = p.get();
+                        auto net = p.get();
                     #else
-                    auto net = p;
+                        auto net = p;
+                        if (!ckpt_load_path.empty()) {
+                            torch::load(p, ckpt_load_path);
+                            model_loaded = true;
+                        }
                     #endif
+
                     params = net->parameters();
                     if (has_cuda) net->to(torch::kCUDA);
-                    LOGGER(INFO) << "Using EfficientNetB0" << std::endl;
+                    LOGGER(INFO) << "Using model: EfficientNetB0" << (model_loaded ? load_str : "") << std::endl;
                     LOGGER(DEBUG) << std::endl << *net << std::endl;
+
                     #ifdef USE_BASE_MODEL
-                    pNet = std::dynamic_pointer_cast<IModel>(p);
+                        pNet = std::dynamic_pointer_cast<IModel>(p);
                     #else
-                    pNet = p;
+                        pNet = p;
                     #endif
                     //pNet = torch::nn::AnyModule(EfficientNetB0(nb_class));
                 }
@@ -302,18 +416,24 @@ int main(int argc, const char* argv[]) {
                     //pNet = std::make_shared<NFNet34>(nb_class);
                     auto p = std::make_shared<NFNet34CLI>(nb_class);
                     #ifdef USE_BASE_MODEL
-                    auto net = p.get();
+                        auto net = p.get();
                     #else
-                    auto net = p;
+                        auto net = p;
+                        if (!ckpt_load_path.empty()) {
+                            torch::load(p, ckpt_load_path);
+                            model_loaded = true;
+                        }
                     #endif
+
                     params = net->parameters();
                     if (has_cuda) net->to(torch::kCUDA);
-                    LOGGER(INFO) << "Using NFNet34" << std::endl;
+                    LOGGER(INFO) << "Using model: NFNet34" << (model_loaded ? load_str : "") << std::endl;
                     LOGGER(DEBUG) << std::endl << *net << std::endl;
+
                     #ifdef USE_BASE_MODEL
-                    pNet = std::dynamic_pointer_cast<IModel>(p);
+                        pNet = std::dynamic_pointer_cast<IModel>(p);
                     #else
-                    pNet = p;
+                        pNet = p;
                     #endif
                     //pNet = torch::nn::AnyModule(NFNet34(nb_class));
                 }
@@ -324,32 +444,84 @@ int main(int argc, const char* argv[]) {
         if (has_cuda) model->to(torch::kCUDA);
         LOGGER(DEBUG) << *model;*/
 
+        /*if (!ckpt_load_path.empty()) {
+            LOGGER(INFO) << "Loading model checkpoint..." << std::endl;
+            #ifdef USE_BASE_MODEL
+            #error Not Implemented checkpoint loading with USE_BASE_MODEL
+            #else
+            torch::load(std::shared_ptr<torch::nn::Module>(pNet.ptr()), ckpt_load_path);
+            #endif
+        }*/
+
+        // TODO: for real resume training, would require to load optimizer param-group states of matching epoch
+        //       for now, only load the checkpoint as is and retrain instead of resume
+
         std::shared_ptr<torch::optim::Optimizer> pOptim;
 
-        switch (optimtype) {
-            case OptimType::Adam:
-                LOGGER(INFO) << "Using Adam" << std::endl;
-                pOptim = std::make_shared<torch::optim::Adam>(params, torch::optim::AdamOptions(lr));
+        switch (optim_type) {
+            case OptimType::Adam: {
+                    auto opt = torch::optim::AdamOptions(lr);
+                    if (lr_opt->count()) opt.lr(lr);
+                    if (betas_opt->count()) opt.betas(betas);
+                    if (epsilon_opt->count()) opt.eps(epsilon);
+                    if (amsgrad_opt->count()) opt.amsgrad(amsgrad);
+                    LOGGER(INFO) << "Using optimizer: Adam" << std::endl
+                        << tab << std::setw(w) << std::left << "Learning Rate" << sep << opt.lr() << std::endl
+                        << tab << std::setw(w) << std::left << "Weight Decay" << sep << opt.weight_decay() << std::endl
+                        << tab << std::setw(w) << std::left << "Betas" << sep
+                            << std::get<0>(opt.betas()) << ", " << std::get<1>(opt.betas()) << std::endl
+                        << tab << std::setw(w) << std::left << "Epsilon" << sep << opt.eps() << std::endl
+                        << tab << std::setw(w) << std::left << "AMSGrad" << sep << opt.amsgrad() << std::endl;
+                    pOptim = std::make_shared<torch::optim::Adam>(params, opt);
+                }
                 break;
-            case OptimType::SGD:
-                LOGGER(INFO) << "Using SGD" << std::endl;
-                pOptim = std::make_shared<torch::optim::SGD>(params, torch::optim::SGDOptions(lr));
+            case OptimType::SGD: {
+                    auto opt = torch::optim::SGDOptions(lr);
+                    if (lr_opt->count()) opt.lr(lr);
+                    if (weight_decay_opt->count()) opt.weight_decay(weight_decay);
+                    if (momentum_opt->count()) opt.momentum(momentum);
+                    if (dampening_opt->count()) opt.dampening(dampening);
+                    if (nesterov_opt->count()) opt.nesterov(nesterov);
+                    LOGGER(INFO) << "Using optimizer: SGD" << std::endl
+                        << tab << std::setw(w) << std::left << "Learning Rate" << sep << opt.lr() << std::endl
+                        << tab << std::setw(w) << std::left << "Weight Decay" << sep << opt.weight_decay() << std::endl
+                        << tab << std::setw(w) << std::left << "Momentum" << sep << opt.momentum() << std::endl
+                        << tab << std::setw(w) << std::left << "Dampening" << sep << opt.dampening() << std::endl
+                        << tab << std::setw(w) << std::left << "Nesterov" << sep << opt.nesterov() << std::endl;
+                    pOptim = std::make_shared<torch::optim::SGD>(params, opt);
+                }
                 break;
-            case OptimType::SGDAGC:
-                LOGGER(INFO) << "Using SGDAGC" << std::endl;
-                pOptim = std::make_shared<torch::optim::SGDAGC>(params, torch::optim::SGDAGCOptions(lr));
+            case OptimType::SGDAGC: {
+                    auto opt = torch::optim::SGDAGCOptions(lr);
+                    if (lr_opt->count()) opt.lr(lr);
+                    if (weight_decay_opt->count()) opt.weight_decay(weight_decay);
+                    if (momentum_opt->count()) opt.momentum(momentum);
+                    if (dampening_opt->count()) opt.dampening(dampening);
+                    if (nesterov_opt->count()) opt.nesterov(nesterov);
+                    if (epsilon_opt->count()) opt.eps(epsilon);
+                    if (clipping_opt->count()) opt.clipping(clipping);
+                    LOGGER(INFO) << "Using optimizer: SGDAGC" << std::endl
+                        << tab << std::setw(w) << std::left << "Learning Rate" << sep << opt.lr() << std::endl
+                        << tab << std::setw(w) << std::left << "Weight Decay" << sep << opt.weight_decay() << std::endl
+                        << tab << std::setw(w) << std::left << "Momentum" << sep << opt.momentum() << std::endl
+                        << tab << std::setw(w) << std::left << "Dampening" << sep << opt.dampening() << std::endl
+                        << tab << std::setw(w) << std::left << "Nesterov" << sep << opt.nesterov() << std::endl
+                        << tab << std::setw(w) << std::left << "Epsilon" << sep << opt.eps() << std::endl
+                        << tab << std::setw(w) << std::left << "Clipping Lambda" << sep << opt.clipping() << std::endl;
+                    pOptim = std::make_shared<torch::optim::SGDAGC>(params, opt);
+                }
                 break;
         }
 
-        //    std::pair<std::vector<std::string>, std::vector<Label>> pairs_training, pairs_validation;
-        //    splitData(pair_images_labels, 0.8, pairs_training, pairs_validation);
+        // DataSamples pairs_training, pairs_validation;
+        // splitData(pair_images_labels, 0.8, pairs_training, pairs_validation);
 
         show_machine_memory();
         show_gpu_memory();
 
         // Initialize DataAugmentationDataset class and read data
         LOGGER(INFO) << "Generating data loaders with data augmentation..." << std::endl;
-        auto dataAugRNG = cv::RNG();
+        auto dataAugRNG = cv::RNG(seed);
         auto custom_dataset_train = DataAugmentationDataset(
             samples_train.first, samples_train.second, image_size, dataAugRNG
         ).map(torch::data::transforms::Stack<>());
@@ -360,8 +532,7 @@ int main(int argc, const char* argv[]) {
         LOGGER(INFO) << "Creating random samplers..." << std::endl;
         using RandomDataLoader = torch::data::samplers::RandomSampler;
         torch::data::DataLoaderOptions loadOpts(batch_size);
-        if (workers >= 0)
-            loadOpts.workers(workers);
+        loadOpts.workers(worker_count);
         auto data_loader_train = torch::data::make_data_loader<RandomDataLoader>(std::move(custom_dataset_train), loadOpts);
         auto data_loader_valid = torch::data::make_data_loader<RandomDataLoader>(std::move(custom_dataset_valid), loadOpts);
         auto train_size = custom_dataset_train.size().value();
@@ -370,7 +541,7 @@ int main(int argc, const char* argv[]) {
         show_gpu_memory();
 
         LOGGER(INFO) << "Starting train/valid loop..." << std::endl;
-        train(pNet, /*lin,*/ data_loader_train, data_loader_valid, pOptim, train_size, valid_size, max_epochs);
+        train(pNet, /*lin,*/ data_loader_train, data_loader_valid, pOptim, train_size, valid_size, max_epochs, ckpt_save_path);
 
         auto end_time = std::chrono::steady_clock::now();
         std::chrono::duration<double> elapsed_seconds = end_time - start_time;